Adding New implementation style
